audithandler: refactor form processor
- Pulls form logic into its own file so handler.go is really just
  handler-related code
- Makes form responsible for applying its filtering to the audit log
  finder
